Implemented kanbanGo write-back mechanism. (#44)
* Minor documentation adjustment.

* ScriptToAsmStr now displays function data in hex.

* Added comments stream for transaction signatures.

* Added signature debug code. Does not pose any security risks.

* Adjusting script->human readable display.

* Staging code for transaction templates.

* Inserted logic for verifying new SCAR transaction format.

Not complete yet, more work to be done.

* PrivateKeyKanban now directly stores secp256k1 data structures, rather than their serializations.

Implemented an ECDSA wrapper in PrivateKeyKanban.

* Test functions for ECDSA<->PrivatekeyKanban binding.

* Bug fix in ToBytesForSignatureWithoutAncestor serialization.

* ECDSA signature in ancestor-less transaction works.

* Bug fix in VerifyFeesCoveredByOneInput. Added Debug code.

* Refactored aggregate signature log handling.

* Refactored logging in ethereum code execution.

* Fee coverage mechanism hooked up. Not tested yet.

* Removed aggregate signature creation from logs.

* Fees covered mechanism, next iteration.

* No-ancestor transaction is now accepted in the mempool.

* Coinbase-like transactions no longer crash the mining process.

* Fees accounted only when non-negative.

* Refactored of mempool to json function. Added debug code.

* More block mining checks cleared.

* Next iteration of accept-fees mining modificaitons.

* Fix test compilation errors. Linker errors still exist.

* Resolved some build issues.

* Renamed coinbaselike->withoutAncestor. Eliminated compiler warnings.

* Attempt to fix qt build.

* Refactored ConnectBlock.

* More detailed error messages in ProcessSmartContract.

* More checks passed in block mining.

* Condensing transaction now appears to work.

* Contract-pays-fees passes through mining.

* Transactions without ancestor pass boot checks in DisconnectBlock().

* Fascstate: Renamed local variables for readability.

* Smart contract balance is now correctly accounted when accepting fees.

* Cleanup of debug code.

* Increased block gas limit. Removed debug code.

* New contractCode rpc call. Added comments stream to EVM executive.

* Added comment stream to EVM.

* Debug code commented out.

* Fix boost tests.

* Reverted EVM logs now provide appropriate feedback.

* Fetch shardId contracts rpc call created. Not implemented yet.

* fetchSCARShardPublicKeys function implemented.

* Added new opcall for scar signature verification.

* Implemented fetchscarshardpublickeys rpc call.

* Towards fetching public keys from EVM scar: SolidityEncodings class implemented.

* FetchSCARShardPublicKeysInternal passed manual testing.

* Cleanup of debug code in block validation code.

* Update of aggregate signature serializations.

* Linker problem: FetchSCARShardPublicKeysInternalPointer added to resolve issues with the build system.

The autogen.sh tools are very difficult to work with.
Proposal: remove autoget.sh for a better build system.

* Fixed defined but not used compiler warning for FetchSCARShardPublicKeysInternalPointer.

* Bootstrapped SCAR signature mechanism. Not tested yet.

* SCAR shardId changed from 32 to 20 bytes.

* Fix broken test build (precomputedtransactiondata name change).

* Void pointer casting to avoid compiler warning: defined but not used.

* Serialization for scar signature now includes scar address and shard id.

* Fixed MakeSCARSignatureTemplate which wasn't working as expected.

* Broken fuzzy test build fixed by commenting out offending code.

* Documentation on the aggregate signature serializations.

* Clean-up of aggregate signature testing APIs, part 1.

Not everything is perfect yet, more clean-up to come.

* Cleanup of aggreagate signature test interfaces, part 2.

* Bug fix in fee acceptance logic and added upper limit to number of log messages displayed on error.

* Adjust transaction checking logic when contract fees are covered by output.

* Clean-up of debug code.

* EVM log messages with bytes in the range [32-126] are now auto-ascii translated.

* Improved evm log messages.

* Attempt to fix broken clang build.

* Added mined block comments size limit.

* Fixed wrong error message in AcceptToMempool.

+++ src/chainparams.h
@@ -88,7 +88,7 @@ protected:
 
     Consensus::Params consensus;
     CMessageHeader::MessageStartChars pchMessageStart;
-    int nDefaultPort;
+    uint16_t nDefaultPort;
     uint64_t nPruneAfterHeight;
     unsigned int nEquihashN = 0;
     unsigned int nEquihashK = 0;
